2001-07-01  Toon Moene  <toon@moene.indiv.nluug.nl>             Pedro Vazquez  <vazquez@penelope.iqm.unicamp.br>
	* libI77/fio.h: Use off_t when appropriate.
	* libI77/backspace.c (f_back): Ditto.
	* libI77/endfile.c (t_runc): Ditto.
	* libI77/err.c (f__nowreading): Ditto.
	* libI77/ftell_.c (unit_chk): Ditto.
	* libI77/sue.c (global f__recloc, s_wsue): Ditto.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@43670 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libf2c/libI77/fio.h
@@ -34,7 +34,7 @@ typedef struct
 	long uinode;
 	int udev;
 #endif
-	int url;	/*0=sequential*/
+	off_t url;	/*0=sequential*/
 	flag useek;	/*true=can backspace, use dir, ...*/
 	flag ufmt;
 	flag urw;	/* (1 for can read) | (2 for can write) */
